wavespeed/requirements.txt
# WaveSpeed AI ComfyUI Nodes Requirements
pydantic>=2.0.0
requests>=2.28.0
Pillow>=9.0.0
urllib3>=1.26.0